Correct: protrusion base curvature
Protrusion stress 3.nb --> Protrusion stress 3-1.nb + relevant figures

Background:
In revising to v3.0, I corrected the equation used to find the curvature of the toroidal base in the direction perpendicular to the tube curvature.

Error:
I used mismatching definitions of the parametric angle defining where you were in the tube from the transverse plane.  The equation pulled from the Kobayashi differential geometry textbook used an angle in which 0 was oriented outward; I had been using an angle in which 0 was oriented inward.  My application of the equation for cuvature was therefore using the wrong angle.

Fix:
Changed angle to be $\pi - \beta$ where \beta is the angle measured from the interior and thus the complement of the angle, $u$, used in the curvature formula.

Impact:
This flipped the sign of kappaT to the correct one and made the mean curvature decrease with z in the base (due to the now opposing signs)