Base fusion before and after: init
Initial attmept to calculate bending energy of the base before and after a fusion event that results in protrusion scission.

Solution 1 using a left Riemann sum moving along the z axis (within the domain of the rim) DID NOT WORK.  There is something weird going on where the computed G_b converges to 0 as the dz decreases, which is definitely wrong.

Solution 2 that directly uses a surface integral seemed complicated, so I didn't finish it yet.

Note: Push done quite late.  This code was done on 2020 Jul 28.