Membrane approach 3
Changed computation to use Kb = 15 kT instead of the previous 10 kT to make it consistent w the Kb value used in my other computations.  This didn't really change anything as far as the results go though.  Also included are the graphs from the computations.  The key for the colors is in 0 Notes.txt

Note: Commit done late.  This was originally done on 2020 Sept 2.